backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
boto3==1.24.66
botocore==1.27.66
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
Django==3.2
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.12.0
drf-yasg==1.21.3
idna==3.3
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
packaging==21.3
pycparser==2.21
PyMySQL==1.0.2
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.2.1
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.12
